lite nya klasser

+++ src/main/java/com/example/pokeblitz/Classes/BattlePokemon.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class BattlePokemon {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String name;
 
     private int maxHp;
@@ -46,7 +46,7 @@ public class BattlePokemon {
     public BattlePokemon() {
     }
 
-    public BattlePokemon(int id, String name) {
+    public BattlePokemon(Long id, String name) {
         Pokemon poke = Client.getPokemonByName(name);
         this.id = id;
         this.name = poke.getName().toUpperCase();
@@ -91,11 +91,11 @@ public class BattlePokemon {
         return returnTypes;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
